chore(gui): show version name on the left menu
